Distinguish between default and user transcoders
Split the TranscoderProfile into a subclass (ExtendedTranscoderProfile), which
has the Id property (as that isn't relevant in the MediaTranscoding namespace)
and a new Type property, which indicates the type: a default (system)
transcoders, which can be updated with every release, or a custom (user)
transcoder.

Save every transcoder that was changed by the user with the type user. This
results in duplicates transcoders with updates, but that's better than data
loss.